refactor: rename module accordingly
