+++ src/types.rs
@@ -1,5 +1,5 @@
-mod opt_arg;
-pub use opt_arg::Opt;
+mod opt;
+pub use opt::Opt;
 mod cruid;
 pub use cruid::Cruid;
 mod engine_time;
